Rename api folder
