+++ apps/api/src/controllers/applicationController.ts
