+++ apps/api/src/routes/application-router.ts
@@ -26,7 +26,7 @@ import {
 	getAllApplications,
 	getApplicationById,
 	getApplicationStateTotals,
-} from '@/api/applicationController.js';
+} from '@/controllers/applicationController.js';
 import { isPositiveNumber } from '@/utils/routes.js';
 
 const applicationRouter = express.Router();
